fix(library): address grip polish review feedback (TASK-30020)
